you can now upload images

+++ src/components/Menu/my_vehicle.js
@@ -7,6 +7,8 @@ import delete_icon from './../Images/delete_icon.svg'
 import edit_icon from './../Images/edit_icon.svg'
 import cancel_button from './../Images/cancelbutton.svg'
 import update_icon from './../Images/update_icon.svg'
+import no_image from './../Images/no_image.png'
+import add_image_icon from './../Images/add_image_icon.svg'
 
 export default class Myvehicle extends Component {
   constructor() {
@@ -20,7 +22,7 @@ export default class Myvehicle extends Component {
       Color: "",
       Size: "",
       Plate: "",
-      car_pic: [],
+      car_pic: '',
       vehicles: [],
       toggle: false,
       edit: false,
@@ -40,10 +42,10 @@ export default class Myvehicle extends Component {
       Model: e.model,
       Color: e.color,
       Size: e.size,
-      Plate: e.plate
+      Plate: e.plate,
+      car_pic: e.car_pic
     })
   }
-  
   handleToggle(){
     this.setState({toggle: !this.state.toggle})
   }
@@ -68,8 +70,10 @@ export default class Myvehicle extends Component {
       plate: this.state.Plate
     }
     axios.post('/api/vehicle', body).then((res) => {
+      this.componentDidMount()
       this.setState({
         vehicles: res.data,
+        toggle: false,
       })
     })
   }
@@ -90,29 +94,87 @@ export default class Myvehicle extends Component {
     })
   }
   deleteCar(id){
-    axios.delete(`/api/vehicle/${id}`).then((res) => {
-      this.componentDidMount()
-    })
+    let result = window.confirm('Are you sure you want to delete this vehicle?')
+    if(result){
+      axios.delete(`/api/vehicle/${id}`).then((res) => {
+        this.componentDidMount()
+      })
+    }
   }
 
   render() {
     let mappedVehicles = this.state.vehicles.map((e, i) => (
+
       <div key ={i}>
         <div style = {{boxShadow: '0px 1px 10px grey', padding: '20px', marginTop: '30px', textAlign: 'left'}}>
-        {e.car_pic ? <img alt = '' src = {e.car_pic} style ={{height: '150px'}}/> : <div style = {{height: '150px', width: '150px', backgroundColor: 'grey'}}></div> }
+
+        {e.car_pic ?
+          <img alt = '' src = {e.car_pic} style ={{height: '150px',margin: 'auto', display:'block'}}/> 
+        : 
+          <div>
+            {this.state.edit ? 
+                <div> 
+                  {/* <label for = 'upload-photo' style = {{curser: 'pointer'}}>
+                    <img alt = ''src = {add_image_icon} style = {{height: '150px',margin: 'auto', display:'block'}}/>
+                  </label>
+                  <input type='file' id = 'upload-photo' accept='image/*' style = {{height: '150px',margin: 'auto', display:'block',display: 'none'}}/> */}
+                </div>
+              :
+                <img alt = '' src = {no_image} style = {{height: '150px', margin: 'auto', display:'block'}}/>
+            }
+          </div>
+        }
+
           <hr/>
-          {this.state.edit ? <p>Year: <input type='' className='input' name = 'Year' onChange = {(e) => {this.handleChange(e)}} value = {this.state.Year} maxLength = '4'/></p> : <p>Year: {e.year}</p>}
+          
+          {this.state.edit ? 
+              <p>Year: <input type='' className='input' name = 'Year' onChange = {(e) => {this.handleChange(e)}} value = {this.state.Year} maxLength = '4'/></p>
+            : 
+              <p>Year: {e.year}</p>
+          }
+
           <hr/>
-          {this.state.edit ? <p>Make: <input type='' className='input' name = 'Make' onChange = {(e) => {this.handleChange(e)}} value = {this.state.Make}/></p> : <p>Make: {e.make}</p>}
+
+          {this.state.edit ?
+             <p>Make: <input type='' className='input' name = 'Make' onChange = {(e) => {this.handleChange(e)}} value = {this.state.Make}/></p> 
+            : 
+             <p>Make: {e.make}</p>
+          }
+
           <hr/>
-          {this.state.edit ? <p>Model: <input type='' className='input' name = 'Model' onChange = {(e) => {this.handleChange(e)}} value = {this.state.Model}/></p> : <p>Model: {e.model}</p>}
+
+          {this.state.edit ?
+              <p>Model: <input type='' className='input' name = 'Model' onChange = {(e) => {this.handleChange(e)}} value = {this.state.Model}/></p> 
+            : 
+              <p>Model: {e.model}</p>
+          }
+
           <hr/>
-          {this.state.edit ? <p>Color: <input type='' className='input' name = 'Color' onChange = {(e) => {this.handleChange(e)}} value = {this.state.Color}/></p> : <p>Color: {e.color}</p>}
+
+          {this.state.edit ?
+              <p>Color: <input type='' className='input' name = 'Color' onChange = {(e) => {this.handleChange(e)}} value = {this.state.Color}/></p> 
+            : 
+              <p>Color: {e.color}</p>
+          }
+
           <hr/>
-          {this.state.edit ? <p>Size: <input type='' className='input' name = 'Size' onChange = {(e) => {this.handleChange(e)}} value = {this.state.Size}/></p> : <p>Size: {e.size}</p>}
+
+          {this.state.edit ?
+              <p>Size: <input type='' className='input' name = 'Size' onChange = {(e) => {this.handleChange(e)}} value = {this.state.Size}/></p> 
+            : 
+              <p>Size: {e.size}</p>
+          }
+
           <hr/>
-          {this.state.edit ? <p>Plate: <input type='' className='input' name = 'Plate' onChange = {(e) => {this.handleChange(e)}} value = {this.state.Plate}/></p> : <p>Plate: {e.plate}</p>}
+
+          {this.state.edit ?
+              <p>Plate: <input type='' className='input' name = 'Plate' onChange = {(e) => {this.handleChange(e)}} value = {this.state.Plate}/></p> 
+            :
+              <p>Plate: {e.plate}</p>
+          }
+
           <hr/>
+
           {this.state.edit ? 
               <img style  = {{height:'30px', float: 'right'}} alt = '' src = {update_icon} onClick = {() => {this.updateCar(e)}}/> 
             :
